+++ src/main/java/com/lipcam/PedidosApiSpring/repositories/ClientesRepository.java
@@ -3,6 +3,6 @@ package com.lipcam.PedidosApiSpring.repositories;
 import com.lipcam.PedidosApiSpring.entities.Clientes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientesRepository extends JpaRepository<Clientes, Integer> {
+public interface ClientesRepository extends JpaRepository<Clientes, Long> {
     Clientes findByCpf(String cpf);
 }
